Title: 자연수 n개의 합, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241024/자연수 n개의 합/sum-of-n-natural-numbers.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int s;
+    long long s;
     cin >> s;
 
     long long left = 1, right = 1e10;
